requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
absl-py=1.0.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
anyio=3.5.0=py37h06a4308_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py37h7f8727e_0
astunparse=1.6.3=pypi_0
attrs=21.4.0=pyhd3eb1b0_0
babel=2.9.1=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.11.1=py37h06a4308_0
bleach=4.1.0=pyhd3eb1b0_0
brotlipy=0.7.0=py37h27cfd23_1003
ca-certificates=2022.3.29=h06a4308_1
cached-property=1.5.2=pypi_0
cachetools=5.0.0=pypi_0
certifi=2021.10.8=py37h06a4308_2
cffi=1.15.0=py37hd667e15_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=7.1.2=pypi_0
cryptography=36.0.0=py37h9ce1e76_0
cycler=0.11.0=pypi_0
debugpy=1.5.1=py37h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.4=py37h06a4308_0
flatbuffers=2.0=pypi_0
fonttools=4.33.2=pypi_0
fuzzy-c-means=1.6.3=pypi_0
gast=0.5.3=pypi_0
google-auth=2.6.6=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.46.0=pypi_0
h5py=3.6.0=pypi_0
hydra-core=1.2.0=pypi_0
idna=3.3=pyhd3eb1b0_0
imageio=2.19.0=pypi_0
imbalanced-learn=0.9.0=pypi_0
importlib-metadata=4.11.3=py37h06a4308_0
importlib_metadata=4.11.3=hd3eb1b0_0
importlib_resources=5.2.0=pyhd3eb1b0_1
imutils=0.5.4=pypi_0
ipykernel=6.9.1=py37h06a4308_0
ipython=7.31.1=py37h06a4308_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
jedi=0.18.1=py37h06a4308_1
jinja2=3.0.3=pyhd3eb1b0_0
joblib=1.1.0=pypi_0
json5=0.9.6=pyhd3eb1b0_0
jsonschema=4.4.0=py37h06a4308_0
jupyter_client=7.2.2=py37h06a4308_0
jupyter_core=4.9.2=py37h06a4308_0
jupyter_server=1.13.5=pyhd3eb1b0_0
jupyterlab=3.3.2=pyhd3eb1b0_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.12.0=py37h06a4308_0
keras=2.8.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.4.2=pypi_0
ld_impl_linux-64=2.35.1=h7274673_9
libclang=14.0.1=pypi_0
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.3.0=hd4cf53a_17
markdown=3.3.6=pypi_0
markupsafe=2.0.1=py37h27cfd23_0
matplotlib=3.5.1=pypi_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mistune=0.8.4=py37h14c3975_1001
nbclassic=0.3.5=pyhd3eb1b0_0
nbclient=0.5.13=py37h06a4308_0
nbconvert=6.4.4=py37h06a4308_0
nbformat=5.3.0=py37h06a4308_0
ncurses=6.3=h7f8727e_2
nest-asyncio=1.5.5=py37h06a4308_0
networkx=2.6.3=pypi_0
notebook=6.4.8=py37h06a4308_0
numpy=1.21.6=pypi_0
oauthlib=3.2.0=pypi_0
omegaconf=2.2.2=pypi_0
opencv-contrib-python=4.5.5.64=pypi_0
opencv-python=4.5.5.64=pypi_0
openssl=1.1.1n=h7f8727e_0
opt-einsum=3.3.0=pypi_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.3.5=pypi_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.1.0=pypi_0
pip=21.2.2=py37h06a4308_0
prometheus_client=0.13.1=pyhd3eb1b0_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
protobuf=3.20.1=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=1.9.0=pypi_0
pygame=2.1.2=pypi_0
pygments=2.11.2=pyhd3eb1b0_0
pylatexenc=2.10=pypi_0
pyopengl=3.1.6=pypi_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyrsistent=0.18.0=py37heee7806_0
pysocks=1.7.1=py37_1
python=3.7.13=h12debd9_0
python-barcode=0.13.1=pypi_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.15.1=pyhd3eb1b0_0
pytz=2022.1=pypi_0
pywavelets=1.3.0=pypi_0
pyyaml=6.0=pypi_0
pyzbar=0.1.9=pypi_0
pyzmq=22.3.0=py37h295c915_2
readline=8.1.2=h7f8727e_1
requests=2.27.1=pyhd3eb1b0_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.8=pypi_0
scikit-image=0.19.2=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.7.3=pypi_0
seaborn=0.11.2=pypi_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=61.2.0=py37h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sklearn=0.0=pypi_0
sniffio=1.2.0=py37h06a4308_1
soupsieve=2.3.1=pyhd3eb1b0_0
sqlite=3.38.2=hc218d9a_0
tabulate=0.8.9=pypi_0
tensorboard=2.8.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow=2.8.0=pypi_0
tensorflow-io-gcs-filesystem=0.25.0=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.13.1=py37h06a4308_0
testpath=0.5.0=pyhd3eb1b0_0
tf-estimator-nightly=2.8.0.dev2021122109=pypi_0
threadpoolctl=3.1.0=pypi_0
tifffile=2021.11.2=pypi_0
tk=8.6.11=h1ccaba5_0
tornado=6.1=py37h27cfd23_0
tqdm=4.64.0=pypi_0
traitlets=5.1.1=pyhd3eb1b0_0
typer=0.3.2=pypi_0
typing-extensions=4.1.1=hd3eb1b0_0
typing_extensions=4.1.1=pyh06a4308_0
urllib3=1.26.9=py37h06a4308_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py37_1
websocket-client=0.58.0=py37h06a4308_4
werkzeug=2.1.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
wrapt=1.14.1=pypi_0
xgboost=1.6.0=pypi_0
xz=5.2.5=h7b6447c_0
zeromq=4.3.4=h2531618_0
zipp=3.7.0=pyhd3eb1b0_0
zlib=1.2.12=h7f8727e_2
